feat: Fix all 20 critical deficiencies — transform architecture from broken to production-ready
P0 Fixes (App-Breaking):
- Created 4 missing shared library files: kafka_client.py, observability.py, middleware.py, api_key.py
- Added BaseModelMixin alias to shared/database.py for backward compatibility
- Added version column to BaseModel for optimistic locking
- Fixed create_db_dependency() to call set_tenant_id() for RLS
- Created missing order-service modules: kafka_handlers.py, payments.py, saga.py
- Fixed order-service main.py (removed HealthStatus/HealthChecker, proper lifespan)
- Fixed admin.py saga calls (replaced get_active_sagas with direct dict access)
- Fixed all 8 new services main.py (proper lifespan, set_dependencies wiring)
- Fixed database.py in 5 services (BaseModelMixin→BaseModel, added ALL_MODELS)

P1 Fixes (Critical Business Logic):
- Float→Numeric for all financial columns across 7 services (15 columns total)
- Compliance service: real FSSAI license validation, GSTIN regex validation, DPDP/RBI checks
- Compliance service: proper GST rate determination (5%/18%), GSTIN validation route
- Pricing service: conflict detection (one surge only), discount cap (50%), time window checks
- Pricing service: condition evaluation, price bounding, rule listing endpoint
- Billing service: real invoice generation with commission/GST calculation, invoice finalization
- Billing service: idempotency check on payouts, Razorpay payout integration
- Payment routes: full Razorpay integration (order creation, HMAC signature verification, webhook, refund)
- Notification dispatcher: real MSG91 SMS, FCM push, WhatsApp Business API, AWS SES email
- Notification: fixed hash() dedup to SHA-256 (deterministic across processes)

Architecture Improvements:
- Kafka client: proper topic resolution, DLQ routing, consumer with manual commit
- Observability: structured JSON logging, OpenTelemetry setup, health check router
- Middleware: tenant extraction, correlation IDs, rate limiting, error handling
- API key auth: secure generation, HMAC hashing, constant-time verification, rotation

+++ services/order-service/app/core/database.py
@@ -19,7 +19,7 @@ from datetime import datetime, timezone
 from sqlalchemy import (
     Column, String, Integer, Float, DateTime, Index, Text,
 )
-from sqlalchemy.dialects.postgresql import UUID as PG_UUID, JSONB
+from sqlalchemy.dialects.postgresql import Numeric, UUID as PG_UUID, JSONB
 from sqlalchemy.orm import Mapped
 
 from shared.utils.database import BaseModel
@@ -69,7 +69,7 @@ class OrderModel(BaseModel):
     eta_risk = Column(String(20), default="LOW", comment="ETA risk level: LOW/MEDIUM/HIGH/CRITICAL")
 
     items = Column(JSONB, default=list, comment="Order items as JSONB array")
-    total_amount = Column(Float, nullable=True, comment="Total order amount in INR")
+    total_amount = Column(Numeric(10, 2), nullable=True, comment="Total order amount in INR")
 
     payment_method = Column(String(30), default="UPI", comment="Payment method: UPI/CARD/WALLET/COD")
     payment_ref = Column(String(100), nullable=True, comment="External payment reference")
@@ -191,7 +191,7 @@ class PaymentModel(BaseModel):
         nullable=True,
         comment="Gateway transaction ID (e.g., Razorpay payment ID)",
     )
-    amount = Column(Float, nullable=False, comment="Payment amount in INR")
+    amount = Column(Numeric(10, 2), nullable=False, comment="Payment amount in INR")
     currency = Column(String(3), default="INR", comment="Currency code")
     status = Column(
         String(20),
